ENS V2 hub updates (#704)
* Updates ENSv2 landing page with new content

Refreshes the ENSv2 landing page with updated copy,
a new applications section, and enhanced graphics to
better communicate the features and benefits of ENSv2.

* Fix type error

* Increase navbar z-index

Ensures navbar remains above other elements on the page.

* Improves ENSv2 page call-to-action button styles

Enhances the visual appearance and spacing of call-to-action buttons on the ENSv2 page and app section for better user experience.

Specifically, adjusts the horizontal padding on smaller screens and introduces letter-spacing to the text for improved readability.

* Ensure consistent file endings

Adds a newline character to the end of the CSS file to ensure consistency across different operating systems and tools.